ISSUE #1610: Radar chart indicating athlete profile
