+++ public/js/app.js
@@ -11,6 +11,7 @@ import MapManager from "./ui/maps";
 import TabsManager from "./ui/tabs";
 import LazyLoad from "../libraries/lazyload.min";
 import DataTableManager from "./ui/data-tables";
+import FullscreenManager from "./fullscreen";
 
 const $main = document.querySelector("main");
 const dataTableStorage = new DataTableStorage();
@@ -25,6 +26,7 @@ const chartManager = new ChartManager(router, dataTableStorage, modalManager);
 const mapManager = new MapManager();
 const tabsManager = new TabsManager(chartManager);
 const dataTableManager = new DataTableManager(dataTableStorage);
+const fullscreenManager = new FullscreenManager(chartManager);
 const lazyLoad = new LazyLoad({
     thresholds: "50px",
     callback_error: (img) => {
@@ -44,6 +46,7 @@ const initElements = (rootNode) => {
     modalManager.init(rootNode);
     chartManager.init(rootNode);
     mapManager.init(rootNode);
+    fullscreenManager.init(rootNode);
 }
 
 modalManager.setInitElements(initElements)
@@ -60,37 +63,6 @@ document.addEventListener('pageWasLoaded', (e) => {
         // Open modal.
         modalManager.open(e.detail.modalId);
     }
-
-    document.querySelectorAll('[data-fullscreen-trigger]').forEach((el) => {
-        el.addEventListener('click', (e) => {
-            e.preventDefault();
-
-            if (document.fullscreenElement) {
-                return;
-            }
-
-            const fullScreenContent = el.closest('[data-fullscreen-content]');
-            fullScreenContent.requestFullscreen().then(() => {
-                chartManager.resizeAll();
-            });
-
-            fullScreenContent.addEventListener('fullscreenchange', () => {
-                el.classList.toggle(
-                    'hidden',
-                    Boolean(document.fullscreenElement)
-                );
-                fullScreenContent.classList.toggle(
-                    'fullscreen-is-enabled',
-                    Boolean(document.fullscreenElement)
-                );
-                fullScreenContent.classList.toggle(
-                    'group',
-                    Boolean(document.fullscreenElement)
-                );
-            });
-
-        });
-    });
 });
 document.addEventListener('pageWasLoaded.heatmap', () => {
     const $heatmapWrapper = document.querySelector('.heatmap-wrapper');
